.travis.yml: initial build matrix
This adds a simple build matrix for building arm and AArch64 builds of
RISU on Travis. It's slightly complicated by having to manually install
the cross toolchain for aarch64 for which is uses the helpful Linaro
toolchain images. This does involve a little messing around to ensure
minimal runtime deps are available.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>