risu: paramterise send/receive functions
This is a precursor to record/playback support. Instead of passing the
socket fd we now pass helper functions for reading/writing and
responding. This will allow us to do the rest of the record/playback
work without hacking up the arch specific stuff.

I've also added a header packet with pc/risu op in it so we can keep
better track of how things are going.

---
v3
  - new for v3
  - arm, aarch64, ppc64