Create new risugen_common module and put some functions in it
Create a new risugen_common module which we can use to store utility
functions that any CPU-specific backend might need to use.
Start it off with the implementations of open_bin(), close_bin(),
insn32() and insn16(), and remove those from the per-CPU files.
This requires us also to abstract out a way for the CPU backend
to specify whether it wants insn32 and insn16 to write big or
little endian values, so create one.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>