eval_with_fields(): Evaluate string in context of calling package
Commit 6a3647ae8918 moved eval_with_fields() into risugen_common.pm,
which broke !constraint and !memory blocks which call functions
defined in the cpu-specific backend packages, because the eval
string is now being evaluated inside the risugen_common package
rather than the risugen_$ARCH package.

Fix this by adding a "package" statement to the code we evaluate
so it is evaluated in the context of the module which calls
eval_with_fields().

The ideal solution to this problem would be to tighten up the
way we evaluate strings from risu files, so that they're run
in their own package which only has imported into it the specific
functions that we want to give them access to. Sadly that is more
work than can really be justified at this point, so we settle
for describing the preferred approach in a comment.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>