configure: Put temporary directory in build tree, and delete it
We were putting the temporary directory we use for configure tests
under /tmp/, but never deleting it, so if you ran configure a lot
you'd build up a big stock of useless directories.

Switch to using a scheme like GNU autotools and QEMU, where we
put the configure tempdir in the build tree, and delete it
just when configure exits successfully.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>