risu: handle trace through stdin/stdout
Trace files can get quite large so it would be useful to be able to
just capture the trace stream with stdin/stdout for processing in a
pipe line. The sort of case where this is useful is for building
static binaries where zlib support is missing for whatever reason.

It can also be used in testing pipelines.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>

---
v6
  - use STDIN/OUT_FILENO instead of fileno(foo)
  - strcmp instead of strncmp