contrib/armasl2risu.py: a script to generate patterns from ASL
ARM publish a description of their architecture in a machine readable
Architecture Specification Language which describes details of the
architecture. For RISU this means we can generate a pattern based on
the ASL rather than manually reading through the PDF and the potential
human error which that entails.

To pick a random example:

  armasl2risu.py --only-desc "half" --insn-suffix "_FP16" *.xml

will generate patterns for all instructions that are half precision
and suffix all the patterns with _FP16.

The --only-desc parameter searches all docvars and can be chained so:

  armasl2risu.py  --only-variant "ARMv8.2" \
    --only-desc "fpsimd" --only-desc "halfprec" *.xml

Will extract only half precision fpsimd instructions.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>