Allow SVE state to be present but empty
Since none of the SVE state is call-saved, the kernel has decided
to zap SVE state across syscalls.  When using memblocks, the first
checkpoint will not have executed an SVE instruction, thus there
is no active SVE state, and thus the SVE state is presented in the
sigcontext as empty.  Correct results for reginfo_init are obtained
by doing nothing, as we have already zeroed the structure.