Split RES_MISMATCH for registers and memory
By remembering the specific comparison that failed, we do not
have to try again when it comes time to report.  This makes
the mem_used flag redundant.  Also, packet_mismatch is now
redundant with RES_BAD_IO.

This means that the only thing that report_match_status does
is to report on register status, so rename to report_mismatch_reg.
Also, we know there is a failure, so don't return a status from
the report.

Signed-off-by: Richard Henderson <richard.henderson@linaro.org>
Message-Id: <20200519025355.4420-14-richard.henderson@linaro.org>